+++ pkg/apis/eventing/register.go
@@ -32,7 +32,7 @@ const (
 	// MTChannelBrokerClassValue is the value we use to specify the
 	// Broker using channels, but the resources (ingress,filter) run
 	// in the system namespace. As in Broker from this repository
-	// pkg/reconciler/mtbroker
+	// pkg/reconciler/broker
 	MTChannelBrokerClassValue = "MTChannelBasedBroker"
 
 	// ScopeAnnotationKey is the annotation key to indicate
