+++ pkg/broker/filter/stats_reporter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ingress
+package filter
 
 import (
 	"context"
@@ -26,25 +26,38 @@ import (
 	"go.opencensus.io/stats"
 	"go.opencensus.io/stats/view"
 	"go.opencensus.io/tag"
-	broker "knative.dev/eventing/pkg/mtbroker"
+	broker "knative.dev/eventing/pkg/broker"
 	"knative.dev/pkg/metrics"
 	"knative.dev/pkg/metrics/metricskey"
 )
 
+const (
+	// anyValue is the default value if the trigger filter attributes are empty.
+	anyValue = "any"
+)
+
 var (
 	// eventCountM is a counter which records the number of events received
-	// by the Broker.
+	// by a Trigger.
 	eventCountM = stats.Int64(
 		"event_count",
-		"Number of events received by a Broker",
+		"Number of events received by a Trigger",
 		stats.UnitDimensionless,
 	)
 
 	// dispatchTimeInMsecM records the time spent dispatching an event to
-	// a Channel, in milliseconds.
+	// a Trigger subscriber, in milliseconds.
 	dispatchTimeInMsecM = stats.Float64(
 		"event_dispatch_latencies",
-		"The time spent dispatching an event to a Channel",
+		"The time spent dispatching an event to a Trigger subscriber",
+		stats.UnitMilliseconds,
+	)
+
+	// processingTimeInMsecM records the time spent between arrival at the Broker
+	// and the delivery to the Trigger subscriber.
+	processingTimeInMsecM = stats.Float64(
+		"event_processing_latencies",
+		"The time spent processing an event before it is dispatched to a Trigger subscriber",
 		stats.UnitMilliseconds,
 	)
 
@@ -53,37 +66,39 @@ var (
 	// go.opencensus.io/tag/validate.go. Currently those restrictions are:
 	// - length between 1 and 255 inclusive
 	// - characters are printable US-ASCII
-	eventTypeKey         = tag.MustNewKey(metricskey.LabelEventType)
+	triggerFilterTypeKey = tag.MustNewKey(metricskey.LabelFilterType)
 	responseCodeKey      = tag.MustNewKey(metricskey.LabelResponseCode)
 	responseCodeClassKey = tag.MustNewKey(metricskey.LabelResponseCodeClass)
 )
 
 type ReportArgs struct {
-	ns        string
-	broker    string
-	eventType string
+	ns         string
+	trigger    string
+	broker     string
+	filterType string
 }
 
 func init() {
 	register()
 }
 
-// StatsReporter defines the interface for sending ingress metrics.
+// StatsReporter defines the interface for sending filter metrics.
 type StatsReporter interface {
 	ReportEventCount(args *ReportArgs, responseCode int) error
 	ReportEventDispatchTime(args *ReportArgs, responseCode int, d time.Duration) error
+	ReportEventProcessingTime(args *ReportArgs, d time.Duration) error
 }
 
 var _ StatsReporter = (*reporter)(nil)
 var emptyContext = context.Background()
 
-// Reporter holds cached metric objects to report ingress metrics.
+// reporter holds cached metric objects to report filter metrics.
 type reporter struct {
 	container  string
 	uniqueName string
 }
 
-// NewStatsReporter creates a reporter that collects and reports ingress metrics.
+// NewStatsReporter creates a reporter that collects and reports filter metrics.
 func NewStatsReporter(container, uniqueName string) StatsReporter {
 	return &reporter{
 		container:  container,
@@ -92,26 +107,25 @@ func NewStatsReporter(container, uniqueName string) StatsReporter {
 }
 
 func register() {
-	tagKeys := []tag.Key{
-		eventTypeKey,
-		responseCodeKey,
-		responseCodeClassKey,
-		broker.ContainerTagKey,
-		broker.UniqueTagKey}
-
 	// Create view to see our measurements.
 	err := metrics.RegisterResourceView(
 		&view.View{
 			Description: eventCountM.Description(),
 			Measure:     eventCountM,
 			Aggregation: view.Count(),
-			TagKeys:     tagKeys,
+			TagKeys:     []tag.Key{triggerFilterTypeKey, responseCodeKey, responseCodeClassKey, broker.UniqueTagKey, broker.ContainerTagKey},
 		},
 		&view.View{
 			Description: dispatchTimeInMsecM.Description(),
 			Measure:     dispatchTimeInMsecM,
-			Aggregation: view.Distribution(metrics.Buckets125(1, 10000)...), // 1, 2, 5, 10, 20, 50, 100, 500, 1000, 5000, 10000
-			TagKeys:     tagKeys,
+			Aggregation: view.Distribution(metrics.Buckets125(1, 10000)...), // 1, 2, 5, 10, 20, 50, 100, 1000, 5000, 10000
+			TagKeys:     []tag.Key{triggerFilterTypeKey, responseCodeKey, responseCodeClassKey, broker.UniqueTagKey, broker.ContainerTagKey},
+		},
+		&view.View{
+			Description: processingTimeInMsecM.Description(),
+			Measure:     processingTimeInMsecM,
+			Aggregation: view.Distribution(metrics.Buckets125(1, 10000)...), // 1, 2, 5, 10, 20, 50, 100, 1000, 5000, 10000
+			TagKeys:     []tag.Key{triggerFilterTypeKey, broker.UniqueTagKey, broker.ContainerTagKey},
 		},
 	)
 	if err != nil {
@@ -121,7 +135,9 @@ func register() {
 
 // ReportEventCount captures the event count.
 func (r *reporter) ReportEventCount(args *ReportArgs, responseCode int) error {
-	ctx, err := r.generateTag(args, responseCode)
+	ctx, err := r.generateTag(args,
+		tag.Insert(responseCodeKey, strconv.Itoa(responseCode)),
+		tag.Insert(responseCodeClassKey, metrics.ResponseCodeClass(responseCode)))
 	if err != nil {
 		return err
 	}
@@ -131,7 +147,9 @@ func (r *reporter) ReportEventCount(args *ReportArgs, responseCode int) error {
 
 // ReportEventDispatchTime captures dispatch times.
 func (r *reporter) ReportEventDispatchTime(args *ReportArgs, responseCode int, d time.Duration) error {
-	ctx, err := r.generateTag(args, responseCode)
+	ctx, err := r.generateTag(args,
+		tag.Insert(responseCodeKey, strconv.Itoa(responseCode)),
+		tag.Insert(responseCodeClassKey, metrics.ResponseCodeClass(responseCode)))
 	if err != nil {
 		return err
 	}
@@ -140,19 +158,41 @@ func (r *reporter) ReportEventDispatchTime(args *ReportArgs, responseCode int, d
 	return nil
 }
 
-func (r *reporter) generateTag(args *ReportArgs, responseCode int) (context.Context, error) {
+// ReportEventProcessingTime captures event processing times.
+func (r *reporter) ReportEventProcessingTime(args *ReportArgs, d time.Duration) error {
+	ctx, err := r.generateTag(args)
+	if err != nil {
+		return err
+	}
+
+	// convert time.Duration in nanoseconds to milliseconds.
+	metrics.Record(ctx, processingTimeInMsecM.M(float64(d/time.Millisecond)))
+	return nil
+}
+
+func (r *reporter) generateTag(args *ReportArgs, tags ...tag.Mutator) (context.Context, error) {
 	ctx := metricskey.WithResource(emptyContext, resource.Resource{
-		Type: metricskey.ResourceTypeKnativeBroker,
+		Type: metricskey.ResourceTypeKnativeTrigger,
 		Labels: map[string]string{
 			metricskey.LabelNamespaceName: args.ns,
 			metricskey.LabelBrokerName:    args.broker,
+			metricskey.LabelTriggerName:   args.trigger,
 		},
 	})
-	return tag.New(
+	// Note that filterType and filterSource can be empty strings, so they need a special treatment.
+	ctx, err := tag.New(
 		ctx,
-		tag.Insert(broker.ContainerTagKey, r.container),
-		tag.Insert(broker.UniqueTagKey, r.uniqueName),
-		tag.Insert(eventTypeKey, args.eventType),
-		tag.Insert(responseCodeKey, strconv.Itoa(responseCode)),
-		tag.Insert(responseCodeClassKey, metrics.ResponseCodeClass(responseCode)))
+		append(tags,
+			tag.Insert(broker.ContainerTagKey, r.container),
+			tag.Insert(broker.UniqueTagKey, r.uniqueName),
+			tag.Insert(triggerFilterTypeKey, valueOrAny(args.filterType)),
+		)...)
+	return ctx, err
+}
+
+func valueOrAny(v string) string {
+	if v != "" {
+		return v
+	}
+	return anyValue
 }
